coord: prevent compaction of indexes during read transactions
When a read transaction starts, hold on to a since_handle for all
indexes in the "time domain" of the first query. This since_handle
will prevent those indexes from compacting past the timestamp of the
transaction. On transaction end, the since_handles are dropped, which
forwards that change to the index Frontiers, and will allow compaction
to proceed.

A "time domain" is a term we are using to mean "other things you might
also want to query". For multiple reads in a single transaction that
all happen at the same time, we need to prevent compaction of anything
that might be read, and we also don't want to force users to
pre-declare everything they will read from. Instead, we pre-emptively
assume that anything else in any schema referenced by the first query
might be later queried, so we prevent compaction of those objects
while the transaction is open.