testdrive: Increase timeouts for FETCH calls
Due to the semantics of the FETCH call, it is excluded from
testdrive's retry logic. Therefore, it is on the test writer
to provide a timeout value that is high enough so that the tests
do not fail even if run in a constrained environment.

This commit makes all the timeout value for FETCH to be 60s
with the exception of those FETCH-es that are not expected
to return any rows, so that the full minute is not waited on those.