mzcompose: hardcode help message
As mentioned in my review of #6515, I think improving the mzcompose help
message is a great idea, but I think we can provide better help if we
just hardcode our optimal message as a string.

The upside to this approach is that we don't need to fight argparse to
get our desired usage string. (E.g., we don't really want to document
the --file option just to say you can't use it.) This commit also adds
docs for the mzcompose-specific commands like "list-compositions" and
"web".