Drop BYO timestamp source on errors (#6616)
Instead of printing error messages to the log (or, perhaps worse,
panic'ing), we should indicate, via error types, that we failed to
parse the transaction message.

Errors are propagated all the way to the top of the timestamper
thread, where any failure to read a message due to parsing and/or
unexpected types will cause the timstamper to drop that source.
Any views which depend on that source will no longer progress.

For now, we're not going to propagate these error messages to
the coordinator.

* Relax restriction on zero-collection transactions

I don't know why our test cases have an example transaction with
zero collections. Instead of changing the test case, let's relax
the error condition in Materialize until I understand that better.

Instead, replace the check with a version that verifies that the
sum of the collection events matches the event_count total.