coord: care about since in determine_frontier
Previously determine_frontier had branching logic that would care about
sinces or uppers based on the presence of an index. I'm not aware of
a reason for this distinction. If there was an index, only the upper
was used in determining a timestamp. We do not guarantee that upper >
since, and so sinks could choose a timestamp < since if the chosen
index hadn't had update_upper run and do its normal "since = upper -
compaction_window" thing. Fix this by consulting sinces and use the
earliest.

Also care about table linearizability while here.

Fixes #6687