debezium-avro: mz_sleep() after configuring Debezium
Due to a race condition in Debezium, tables that are created
immediately after configuring Debezium are not detected and replicated.

They are not made part of the initial snapshot and their presence
is not detected afterwards. No schema registry entries are created
which in turn causes Mz to error out.