coord: prevent ship_dataflow from erroring
Previously if ship_dataflow errored (the only way was for the as_of to
be < since, but I found this due to another commit that would have added
another error path), it would leave the catalog in an unsafe state. The
catalog entries would exist with no index, and materialized would crash
when attempting to use the indexes. This happened because catalog_transact
is almost always called before ship_dataflow.

Remove the error return from ship_dataflow to prevent this. The only
current error is due to a bad as_of, but I'm not aware of a way to
trigger that path.