kafka: add consistency_topic to mz_kafka_sinks system table
The main motivation is to be able to use this in testdrive to fetch the
actual consistency topic, instead of relying on a fixed naming scheme
for the topic. We need this because one of the next commits makes the
consistency_topic user configurable.

Additionally, it's good to have this for users.