Add nightly Redpanda job to CI
Add a nightly CI job that will run all Kafka-related tests aginst
Redpanda, except for those that involve dynamically adding partitions to
Kafka topics, as that is not yet supported by Redpanda.

Why a nightly job, and not a per-PR job? Running these tests on every PR
will slow down CI and potentially block PRs that hit bugs in Redpanda.
That's a tradeoff worth making when we're ready to publicly commit to
Redpanda support, but I don't think we're quite ready yet. We're now
well positioned to make that commitment in the future, though.

Note that @cirego, @senior7515, and the rest of the Redpanda team did
the heavy lifting to make Materialize and Redpanda work together. See
the many comments on #6078 for the details on that work.

Closes #6078.
Closes #6800.

Co-authored-by: Chris Golden <cirego@gmail.com>
Co-authored-by: Alexander Gallego <alex@vectorized.io>