persist: implement Persister backed by a Buffer and Blob
The bulk of the logic is handled by the new single-threaded `Indexed`
data structure, which wraps a `Buffer` (for quickly persisting new
data), a `BlobFuture` which indexes unsealed data by time, and a
`BlobTrace` which indexes data once it has been sealed. Matching this to
Persister's concurrency of multiple streams is very much still a work in
progress.