Add graceful retries for the S3 source
Up to now, the S3 source could tolerate failures when receiving SQS
notifications, but it could not handle failures when accessing objects,
both for notifications and scans. This means that event transitory
problems would cause the source to stop responding to updated data.

This change creates the foundation to let the asynchronous tasks within
the s3 module communicate detailed error information and retry
conditions.

Currently, the retry policy is to retry ten times, with a delay of ten
seconds. This should be replaced by a back-off logic, preferably
configurable.

The change does not include tests, which will need to be addressed
separately due to the integration complexity.

Issue: #6010

Signed-off-by: Moritz Hoffmann <mh@materialize.com>