ci: always trigger a build
relying on downstream jobs to depend on the build step doesn't always do
the right thing and can lead to broken code to be merged in main like in
MaterializeInc/materialize#6910.

Since pretty much every single substep relies on the build step, this PR
makes the build step trigger with any file change.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>