kafka: add kafka_time_offset option
The kafka_time_offset option will be translated to the existing
start_offset option during purification. After lookup of the offsets for
the provided timestamp, things behave as if start_offset was specified
directly. You can think of kafka_time_offset as a convenient way of
providing start_offset when you don't care about the exact start_offset
value or you can somehow correlate time with offsets easily.

For each partition, the set offset is the earliest offset with a
timestamp greater than or equal to the requested timestamp. If no such
offset exists for a partition (e.g. kafka_time_offset is greater than
the greatest timestamp in the partition), we start from the end (by
looking up the current high watermark of the partition). If a partition
is added during runtime, we start from the beginning (which is what
start_offset does if no offset is provided for a new partition).

Note that we expect to lift these from WITH options to become regular
columns of the source which is the more general solution and also more
idiomatic.