doc/design: add proposal for one-shot sinks
Some of our users might want and easy way of getting the current data in
a view out of Materialize, without worrying about a continuous stream of
changes, exactly once, and changing topic names or file names. This is
at odds with how current sinks work and at odds with potential stricter
requirements of what I'm retroactively calling _continuous sinks_.

I propose to add the concept of _one-off sinks_, which only write out
data up to a point and then finish. They should have relaxed
requirements and should be easier to implement in the future.