persist: start documenting and debug_asserting some invariants
Some of these are just sanity checks and some are going to be necessary
to provide indexing on top of the persisted batches. Get this started
early to get in the habit of thinking about what our invariants are.

Also sneak in a file-based implementation of Blob since it's
straightforward and this PR includes a generic set of assertions to run
over anything implementing Blob.s