sinks,avro-ocf: create Avro Writer/open File only when needed
All the writing of a Avro OCF sink happens from one worker. However,
before this change all workers would open a file, with the same file
name even.

This changes two things:
 - We make it explicit that we route all messages to one worker
 - We lazily create the Avro Writer when we receive the first messages

The previous code would silently "continue" with the sink, only logging
and error once, when creating the writer failed. Now, we log an error
every time we try and create the writer, which will make the error more
noticeable in the logs.