sql-parser: fix parsing of dollar-quoted strings
The lexer was not correctly skipping over the second `$` character in
the opening tag. This bug has been present since the lexer was rewritten
with support for this style of string literal, and was reflected in the
test cases already... just no one had noticed until now (h/t @ruchirK).

Fix #7198.