demo/debezium: remove
The debezium demo does not meet the standard for officially supported
demos described in doc/developer/demos.md. The demo could plausibly live
in play, but it brings in a lot of duplicate dependencies (see the
massive red diff in Cargo.lock), so for now I think it's best if we
resurrect these tools on an as-needed outside of the main repository!