coord: remove table granularity logging
Previously tables would check to see if they needed a manual advancing
(not triggered by `need_advance`) on each Coordinator message. While
not expensive to performance, it was confusing to know exactly what was
going on and why.

After some investigation and discussion, it seems that the original reason
for this is perhaps gone. When tables were added, there were no internal
system tables, and so something needed to advance table timestamps if
there were no table operations, which led to the logging_granularity
stuff. Now there are some system tables that are updated each second,
rendering this manual update needless (I think). Thus it seems that we
can remove the manual table advance.

If we discover some negative user experience or flakey tests then we
can add this back, but with increased understanding of why it's needed,
with improved documentation or testing to assert the behavior we need.

The comment on `logging_granularity` was for a different variable
previously removed and should never have been attached to
`logging_granularity`.