coord: maintain at least one timestamp binding that is not beyond the write frontier
Before, we were compacting away any binding that was not beyond the
frontier. This means that on restart, it can happen that previously
written data gets assigned a timestamp that is beyond the previous write
frontier, meaning it would get written out again.

Now, we maintain the largest binding that is not beyond the frontier to
ensure that previously written data gets a timestamp that is not beyond
the write frontier when restarting.