tests: simplify kafka exactly-once test, make it properly detect issues
Before, relying on reingesting the RT topic could give false results if
reingestion was too slow.

Now, we straight up verify the output topic of the real-time
timestamping pipeline.

We also wait before restarting, to verify correct results even with
timestamp binding compaction. For this, we also have to set the
compaction window to 1ms.

This also removes the SELECT statements which were not really doing
anything.