dataflow,sinks: continuously write out END records when updating write frontier
When restarting, the Kafka sink determines the gate timestamp (the
timestamp used to filter out previously written records) based on the
END records in the consistency topic.

We therefore need to ensure that the latest timestamp in those records
advances in lockstep with the write frontier, which is used to compact
timestamp bindings.