persist: fix non-sequential future batches bug
This is one of the invariants we maintain (for now anyway, neither
ruchirk nor I are particularly convinced at the moment that this
invariant is buying us much). On the other hand, this commit fixes it in
a straightforward and lightweight way, so may as well keep it for now.

The bug comes in two flavors, both fixed in this commit.

First is some stream is registered, written to, and step'd, moving seqno
0..X into future. Then a second stream is registered, written to, and
step'd. When it goes to move X..Y into the future, the second stream is
missing a batch for 0..X. (Newly registered streams are missing 0 to the
seqno that buffer was at when they are registered.)

The second flavor is similar. If we then write to the first stream again
and step, it is then missing X..Y. (A stream not written to between two
step calls doesn't get a batch.)

This also has the nice side effect of updating META once per drain_buf
instead of once per stream written to in the drained records.

Found by nemesis.

This plus ruchirk's recent consolidation PR fix all the outstanding bugs
caught by nemesis, so un-ignore the test.