s3 source: Correct exits from download_objects loop
Only break out of the loop/exit the task if the downstream (actual dataflow)
task has shut down, not every time there is an error -- some, and ideally
eventually all -- errors are recoverable in some way.

Every time we exit the loop we should explicitly close the receiver so that
upstream tasks don't wait an unbounded amount of time to observe the closure
being dropped. This is more of hygiene than a correctness thing.